fix(backend): Refactor grid and payment logic to service layer
